+++ src/orquestra/sdk/_ray/_ray_logs.py
@@ -15,6 +15,7 @@ from pathlib import Path
 
 import pydantic
 
+from orquestra.sdk.schema.ir import TaskInvocationId
 from orquestra.sdk.schema.workflow_run import TaskRunId, WorkflowRunId
 
 from . import _client
@@ -50,6 +51,8 @@ class WFLog(pydantic.BaseModel):
     message: str
     # ID of the workflow that was run when this line was produced.
     wf_run_id: t.Optional[WorkflowRunId]
+    # ID of the task invocation (node inside the workflow def graph).
+    task_inv_id: t.Optional[TaskInvocationId]
     # ID of the task that was run when this line was produced.
     task_run_id: t.Optional[TaskRunId]
 
@@ -144,12 +147,6 @@ class _RayLogs:
 
         return logs
 
-    def iter_logs(self):
-        while True:
-            parsed_logs = self._read_log_files()
-            yield [log.json() for log in parsed_logs]
-            time.sleep(2)
-
     def get_full_logs(self):
         parsed_logs = self._read_log_files()
         formatted = [log.json() for log in parsed_logs]
@@ -173,7 +170,3 @@ class DirectRayReader:
     def get_full_logs(self, run_id: t.Optional[str] = None) -> t.Dict[str, t.List[str]]:
         wrapped = _RayLogs(ray_temp=self._ray_temp, workflow_or_task_run_id=run_id)
         return wrapped.get_full_logs()
-
-    def iter_logs(self, run_id: t.Optional[str] = None) -> t.Iterator[t.Sequence[str]]:
-        wrapped = _RayLogs(ray_temp=self._ray_temp, workflow_or_task_run_id=run_id)
-        yield from wrapped.iter_logs()